requirements.txt
ollama==0.6.1
openai==2.15.0
pillow==12.1.0
pip==25.3
psutil==7.2.1
